refinement type
